requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langgraph
pandas
numpy
plotly
altair
python-dotenv
duckduckgo-search>=5.0
fpdf2
ddgs
